move rwlock to separate module
